RDKCOM-3893: update data type ( ULONG -> uint32_t ) + misc code cleanup
Reason for change:
update CosaUtilGetIfAddr() result data type ( ULONG -> uint32_t ) + misc code cleanup
An IPv4 address is a 32bit value.
Test Procedure: Sanity.
Risks: None.

(cherry picked from commit fce23cd73874b4e28affd6c8ebaa74e447e691f2)

+++ source/TR-181/middle_layer_src/cosa_ip_dml.c
@@ -2515,7 +2515,7 @@ IPv4Address_GetParamUlongValue
 #endif /* !_COSA_BCM_MIPS_ */
         {
             /* TBC -- this should be from a SBAPI call */
-            *puLong = CosaUtilGetIfAddr(pIPInterface->Cfg.LinkName);
+            *puLong = (ULONG) CosaUtilGetIfAddr(pIPInterface->Cfg.LinkName);
         }
          
          return TRUE;
